+++ cpp/tests/interop/to_arrow_test.cpp
@@ -97,6 +97,12 @@ std::pair<std::unique_ptr<cudf::table>, std::shared_ptr<arrow::Table>> get_table
 struct ToArrowTest : public cudf::test::BaseFixture {
 };
 
+template <typename T>
+struct ToArrowTestDurationsTest : public cudf::test::BaseFixture {
+};
+
+TYPED_TEST_CASE(ToArrowTestDurationsTest, cudf::test::DurationTypes);
+
 TEST_F(ToArrowTest, EmptyTable)
 {
   auto tables = get_tables(0);
@@ -123,6 +129,39 @@ TEST_F(ToArrowTest, DateTimeTable)
                                             arrow::default_memory_pool());
   timestamp_builder.AppendValues(std::vector<int64_t>{1, 2, 3, 4, 5, 6});
   CUDF_EXPECTS(timestamp_builder.Finish(&arr).ok(), "Failed to build array");
+
+  std::vector<std::shared_ptr<arrow::Field>> schema_vector({arrow::field("a", arr->type())});
+  auto schema = std::make_shared<arrow::Schema>(schema_vector);
+
+  auto expected_arrow_table = arrow::Table::Make(schema, {arr});
+
+  auto got_arrow_table = cudf::to_arrow(input_view, {"a"});
+
+  ASSERT_EQ(expected_arrow_table->Equals(*got_arrow_table, true), true);
+}
+
+TYPED_TEST(ToArrowTestDurationsTest, DurationTable)
+{
+  using T = TypeParam;
+
+  auto data = {T{1}, T{2}, T{3}, T{4}, T{5}, T{6}};
+  auto col  = cudf::test::fixed_width_column_wrapper<T>(data);
+
+  cudf::table_view input_view({col});
+
+  std::shared_ptr<arrow::Array> arr;
+  arrow::TimeUnit::type arrow_unit;
+  switch (cudf::type_to_id<TypeParam>()) {
+    case cudf::type_id::DURATION_SECONDS: arrow_unit = arrow::TimeUnit::type::SECOND; break;
+    case cudf::type_id::DURATION_MILLISECONDS: arrow_unit = arrow::TimeUnit::type::MILLI; break;
+    case cudf::type_id::DURATION_MICROSECONDS: arrow_unit = arrow::TimeUnit::type::MICRO; break;
+    case cudf::type_id::DURATION_NANOSECONDS: arrow_unit = arrow::TimeUnit::type::NANO; break;
+    case cudf::type_id::DURATION_DAYS: return;
+    default: CUDF_FAIL("Unsupported duration unit in arrow");
+  }
+  arrow::DurationBuilder duration_builder(duration(arrow_unit), arrow::default_memory_pool());
+  duration_builder.AppendValues(std::vector<int64_t>{1, 2, 3, 4, 5, 6});
+  CUDF_EXPECTS(duration_builder.Finish(&arr).ok(), "Failed to build array");
 
   std::vector<std::shared_ptr<arrow::Field>> schema_vector({arrow::field("a", arr->type())});
   auto schema = std::make_shared<arrow::Schema>(schema_vector);