+++ python/cudf/cudf/_lib/string_casting.pyx
@@ -41,6 +41,10 @@ from cudf._lib.cpp.strings.convert.convert_urls cimport (
     url_encode as cpp_url_encode,
     url_decode as cpp_url_decode
 )
+from cudf._lib.cpp.strings.convert.convert_durations cimport (
+    to_durations as cpp_to_durations,
+    from_durations as cpp_from_durations
+)
 from cudf._lib.cpp.types cimport (
     type_id,
     data_type,
@@ -565,6 +569,73 @@ def timestamp2int(
                 input_column_view,
                 out_type,
                 c_timestamp_format))
+
+    return Column.from_unique_ptr(move(c_result))
+
+
+def timedelta2int(
+        Column input_col,
+        **kwargs):
+    """
+    Converting/Casting input string column to TimeDelta column with specified
+    format
+
+    Parameters
+    ----------
+    input_col : input column of type string
+
+    Returns
+    -------
+    A Column with string represented in TimeDelta format
+
+    """
+    if input_col.size == 0:
+        return as_column([], dtype=kwargs.get('dtype'))
+    cdef column_view input_column_view = input_col.view()
+    cdef type_id tid = <type_id> (
+        <underlying_type_t_type_id> (
+            np_to_cudf_types[kwargs.get('dtype')]
+        )
+    )
+    cdef data_type out_type = data_type(tid)
+    cdef string c_duration_format = kwargs.get('format').encode('UTF-8')
+    cdef unique_ptr[column] c_result
+    with nogil:
+        c_result = move(
+            cpp_to_durations(
+                input_column_view,
+                out_type,
+                c_duration_format))
+
+    return Column.from_unique_ptr(move(c_result))
+
+
+def int2timedelta(
+        Column input_col,
+        **kwargs):
+    """
+    Converting/Casting input Timedelta column to string
+    column with specified format
+
+    Parameters
+    ----------
+    input_col : input column of type Timedelta in integer format
+
+    Returns
+    -------
+    A Column with Timedelta represented in string format
+
+    """
+
+    cdef column_view input_column_view = input_col.view()
+    cdef string c_duration_format = kwargs.get(
+        'format', "%D days %H:%M:%S").encode('UTF-8')
+    cdef unique_ptr[column] c_result
+    with nogil:
+        c_result = move(
+            cpp_from_durations(
+                input_column_view,
+                c_duration_format))
 
     return Column.from_unique_ptr(move(c_result))